Update format test references to Typst v0.13
Typst v0.13 slightly reduced the widths of single alphabetic characters in attachments and fractions. In older versions of Typst the width would change between a single character and multiple characters. The current behavior is therefore more consistent!
The other cases have a slightly reduced padding between the base unit and the exponent. This is however barely noticeable compared to the first change.
I think these changes are related to https://github.com/typst/typst/issues/274 and https://github.com/typst/typst/pull/5738. While the output will be slightly different in older versions of Typst, this is overall not a huge issue.