Make use of context conditional
The context will only be used now if one of the function arguments is auto. In that case the context is used to read the configuration (or macros) from the state. If all arguments are explicitly passed to the functions, the context is not necessary and can therefore be skipped.